Fix score calculation with long segments overlapping centromeres
The old code tried to account for this, but failed to handle the
initial join where we would get duplicated segments. This caused
subtle bugs which would not manifest if segments did not span
chromosomes (and thus would not trigger the faulty join() behavior)

As a solution, clean up the data after the join, and normalize
against the whole chromosome arm instead, if this happens.